Move users to apps #2
